requirements.txt
django-mediamosa
django-tinymce